Merge

+++ Game/Logic/ActionWorker.cs
@@ -396,7 +396,7 @@ namespace Game.Logic {
         public PassiveAction FindAction(GameObject workerObject, Type type) {
             return passive.Values.FirstOrDefault(action => action.WorkerObject == workerObject && action.GetType() == type);
         }
-
+        
         public void Remove(GameObject workerObject, ActionInterrupt actionInterrupt, params GameAction[] ignoreActions) {
             List<GameAction> ignoreActionList = new List<GameAction>(ignoreActions);
                         
@@ -454,7 +454,7 @@ namespace Game.Logic {
             ((PassiveAction)item).UserCancelled();
         }        
 
-        public Error Cancel(ushort id) {
+        public Error Cancel(uint id) {
             ActiveAction activeAction;
             if (ActiveActions.TryGetValue(id, out activeAction) && !activeAction.isDone) {
                 if ((ActionFactory.GetActionRequirementRecord(activeAction.WorkerType).list[activeAction.WorkerIndex-1].option & ActionOption.UNCANCELABLE) == ActionOption.UNCANCELABLE) {
